fix: renaming LinkedList to SinglyLinkedList
